% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bib2yaml.R
\name{bib2yaml}
\alias{bib2yaml}
\title{Convert a BibTeX to a YAML file}
\usage{
bib2yaml(data = NULL, file = NULL, path = ".", write = TRUE)
}
\arguments{
\item{data}{A vector of collapsed BibTeX [strings] (i.e. each element of the vector/list is one single BibTeX)}

\item{file}{A vector of BibTeX files to be converted}

\item{path}{The path to the directory where YAML files will be written (ignored if \code{write = FALSE})}

\item{write}{A boolean indicating if YAML will be written}
}
\value{
A list of YAML files (i.e. a tagged list). If \code{write = TRUE} YAML files are also written in the directory specified by the argument \code{path}.
}
\description{
This function converts one (or many) BibTeX into YAML format using the software pandoc and pandoc-citeproc (called with the R function\code{system()}).
}
\details{
User has to specify one of \code{data} or \code{file} otherwise the function will return an error.
}
\examples{
# Coming soon...
}
\seealso{
\code{\link[refR]{getRefs}}
}
