% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanRefs.R
\name{cleanRefs}
\alias{cleanRefs}
\title{Clean references.}
\usage{
cleanRefs(files)
}
\arguments{
\item{files}{Filename(s) of references in the YAML format to be cleaned}
}
\description{
Remove LaTeX tags and shorten authors names in YAML references
}
\details{
The conversion from BibTeX to YAML (using pandoc-citeproc called by
\code{\link[refR]{getRefs}} or \code{\link[refR]{bib2yaml}}) removes a lot of
LaTeX tags. But this process is not perfect. The function \code{cleanRefs}
cleans the remaining special characters ('exotic' accents and other LaTeX).
Moreover information in BibTeX is not uniform for authors, so this function
also clean author names (detection of family names with particle, removal of
uppercase except for each first letter of family and given names, and shorten
given name). Note that journal, booktitle and article/chapter title are also
not uniform but this information comes from Scopus that is more standardize
than Crossref.
Finally we have to mention that cleaned YAML files erase their previous versions
(this function does not return any object).
If errors occur and if you still have BibTeX, you can use the function
\code{\link[refR]{bib2yaml}} to regenerate the erased YAML file.
}
\examples{
# Coming soon...
}
\seealso{
\code{\link[refR]{getRefs}}, \code{\link[refR]{bib2yaml}}
}
