% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectBin.R
\name{detectBin}
\alias{detectBin}
\title{Detect a command binary}
\usage{
detectBin(bin)
}
\arguments{
\item{bin}{a system command to be detected, as a character string.}
}
\value{
The path to the command binary.
}
\description{
\code{detectBin} detects and returns the path to a command binary.
}
\examples{
detectBin('cd')
}
