% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractBib.R
\name{extractBib}
\alias{extractBib}
\title{Substract Bibtex entries.}
\usage{
extractBib(bib, text, con, markdown = FALSE)
}
\arguments{
\item{bib}{an object of class \code{BibEntry} or \code{bibentry}.}

\item{text}{a character string to be searched within.}

\item{con}{a connection object or a character string that stand for a path,
ignored if text is defined.}

\item{markdown}{a logical. If \code{TRUE}, Markdown formatted citation marks are sought.}
}
\description{
Substarct Bibtex entries that are encountered in a given text.
}
\examples{
file.name <- system.file('Bib', 'RJC.bib', package='RefManageR')
bib <- RefManageR::ReadBib(file.name)
# bib[[200]]
txt <- 'First \\\\cite{kim1995estimating}, second \\\\Citep{fu2006statistical}'
out <- extractBib(bib, txt)
# RefManageR::WriteBib(out, file=tempfile(fileext = '.bib'))
}
