% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findRef.R
\name{findRef}
\alias{findRef}
\title{Find citation tags.}
\usage{
findRef(text, con, markdown = FALSE)
}
\arguments{
\item{text}{a character string.}

\item{con}{a connection object or a character string that stands for a path,
ignored if \code{text} is used.}

\item{markdown}{a logical. If \code{TRUE},  Markdown's citation tags are sought.}
}
\value{
A data frame with the citation keys and their frequency.
}
\description{
Finds Markdown and Latex citation tags in a character string or a file.
}
\examples{
test <- findRef(text='First \\\\cite{Pimm2000}, second \\\\Citep{May1972}')
}
