% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRefs.R
\name{getRefs}
\alias{getRefs}
\title{Retrieve author references from Scopus and Crossref APIs}
\usage{
getRefs(
  api_key = NULL,
  author_id = NULL,
  date = NULL,
  sort = "pubyear",
  sleep = 5,
  folder = ".",
  format = c("bibtex", "yaml"),
  erase = FALSE
)
}
\arguments{
\item{api_key}{A string indicating the user Scopus API Key obtained from the Elsevier Developer Portal}

\item{author_id}{A string indicating the Scopus identifier of the author for which references have to be retrieve}

\item{date}{A string indicating the year (or a range of years) for which references have to be retrieve (if \code{NULL}, default, all references listed in Scopus for this author will be returned)}

\item{sort}{A string indicating the field by which references will be ordered (one of \code{artnum}, \code{citedby-count}, \code{publicationName}, \code{pubyear})}

\item{sleep}{A positive numeric indicating the time interval (in seconds) between two consecutive Scopus queries (a Scopus query response is limited to 25 entries)}

\item{folder}{A string indicating the folder (relative or absolute path) to write references}

\item{format}{A string indicating the references format (the possible formats are \code{bibtex}, \code{yaml} and \code{json})}

\item{erase}{A boolean specifying if a reference that already exists in the folder must be rewritten (\code{TRUE}) or not (\code{FALSE})}
}
\value{
A list of BibTeX strings. BibTeX (if required) and other specified formats are also written in the directory specified by the argument \code{folder}.
}
\description{
This function queries the Scopus Search API to returned a list of references for one specific author.
Quality of metadata is increased by coupling Scopus results with a query of
the Crossref API called using the package \code{\link[rcrossref]{rcrossref}}.
Results are returned as BibTeX files (one per reference) but user can also choose two other formats: YAML and JSON.
The conversion is proceeded using pandoc and pandoc-citeproc (must be separately installed).
}
\details{
Before using this function user has to get a free Scopus API Key from the Elsevier Developer Portal (\url{https://dev.elsevier.com/user/registration}).
He also needs to know the author Scopus identifer by visiting the Scopus Author Search Portal (\url{https://www.scopus.com/freelookup/form/author.uri}).
This function works in four steps:
\enumerate{
\item Get a list of references from Scopus API and keep only those with a DOI (all fields are returned except the authors list);
\item Complete the metadata (mainly the authors list) using the Crossref API called by the function \code{\link[rcrossref]{cr_cn}} (request by DOI);
\item Merge metadata from the two sources and keep the best (hopefully);
\item Write references (one file per reference) in the specified formats using pandoc-citeproc (for YAML and JSON).
We strongly recommend to convert references in YAML (and also keep BibTeX formats)
because this step (performed with pandoc-citeproc) translates a large amount of LaTeX tags (mainly accented characters).
But if you prefer you can also export references in BibTeX and then use the function \code{\link[refR]{bib2yaml}} to convert them in YAML.
}
}
\examples{
# Coming soon...
}
\seealso{
\code{\link[refR]{bib2yaml}}, \code{\link[refR]{cleanRefs}}
}
